# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ГОСТИННИЦЫ - API",
        "description": "Описание всех методов приложения ГОСТИННИЦЫ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Авторизация в системе",
                "description": "Авторизация в системе",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "email": "test@mail.ru",
                                    "password": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Авторизация прошла успешно",
                                        "token": "abracadabra",
                                        "id": 1,
                                        "first_name": "",
                                        "last_name": "",
                                        "second_name": "",
                                        "email": "test@mail.ru"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registration": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Регистрация в системе",
                "description": "Регистрация в системе",
                "operationId": "registration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "id": 1,
                                    "first_name": "",
                                    "last_name": "",
                                    "second_name": "",
                                    "email": "test@mail.ru",
                                    "password": "123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Авторизация прошла успешно",
                                        "token": "abracadabra",
                                        "id": 1,
                                        "first_name": "",
                                        "last_name": "",
                                        "second_name": "",
                                        "email": "test@mail.ru"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/hotels/getById": {
            "get": {
                "tags": [
                    "Отели"
                ],
                "summary": "Получить объект отеля по ID",
                "description": "Получить объект отеля по ID",
                "operationId": "showHotels",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "ID отеля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": "1",
                                    "name": "Название отеля",
                                    "address": "г. Нижний Тагил, ул. Мира 56, д. 2",
                                    "description": "Супер-дупер отель",
                                    "user_id": "1",
                                    "geo_lat": "44.222",
                                    "geo_lon": "44.222",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotels/store": {
            "post": {
                "tags": [
                    "Отели"
                ],
                "summary": "Создание нового объекта отеля",
                "description": "Создание нового объекта отеля",
                "operationId": "storeHotels",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Название отеля",
                                    "address": "г. Нижний Тагил, ул. Мира 56, д. 2",
                                    "description": "Супер-дупер отель",
                                    "user_id": "1",
                                    "geo_lat": "44.222",
                                    "geo_lon": "44.222",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "name": "Название отеля",
                                        "address": "г. Нижний Тагил, ул. Мира 56, д. 2",
                                        "description": "Супер-дупер отель",
                                        "user_id": "1",
                                        "geo_lat": "44.222",
                                        "geo_lon": "44.222",
                                        "deleted_at": "2023-07-12 06:45:23",
                                        "created_at": "2023-07-12 06:45:23",
                                        "updated_at": "2023-07-12 06:45:23"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotels/update": {
            "put": {
                "tags": [
                    "Отели"
                ],
                "summary": "Обновление объекта отеля",
                "description": "Обновление объекта отеля",
                "operationId": "updateHotels",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "ID отеля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Название отеля",
                                    "address": "г. Нижний Тагил, ул. Мира 56, д. 2",
                                    "description": "Супер-дупер отель",
                                    "user_id": "1",
                                    "geo_lat": "44.222",
                                    "geo_lon": "44.222",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "name": "Название отеля",
                                        "address": "г. Нижний Тагил, ул. Мира 56, д. 2",
                                        "description": "Супер-дупер отель",
                                        "user_id": "1",
                                        "geo_lat": "44.222",
                                        "geo_lon": "44.222",
                                        "deleted_at": "2023-07-12 06:45:23",
                                        "created_at": "2023-07-12 06:45:23",
                                        "updated_at": "2023-07-12 06:45:23"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotels/delete": {
            "delete": {
                "tags": [
                    "Отели"
                ],
                "summary": "Удаление объекта отеля по ID",
                "description": "Удаление объекта отеля по ID",
                "operationId": "destroyHotels",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "ID отеля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}